==> R/00_all.qmd <==
---
title: "Complete Project Documentation"
format:
  revealjs:
    embed-resources: true
editor: visual
---

```{r}
#| label: 00-01

# Loading library
library(quarto)
library(knitr)
library(rmarkdown)
```

{{< include 01_dat_load.qmd >}}

{{< include 02_dat_clean.qmd >}}

{{< include 03_augment.qmd >}}

{{< include 04_describe.qmd >}}

{{< include 05_analysis_1.qmd >}}

```{r}
#| label: 00-02

# Render the presentation separately to the 'doc' directory
quarto_render(input = "./../doc/presentation.qmd")

```
