==> R/01_dat_load.qmd <==
---
title: "Data Loading"
format:
  revealjs:
    embed-resources: true
editor: visual
---

# Data Loading

## Importing packages and libraries

In this section, we begin by importing essential R packages that will be used throughout the data loading process.

```{r}
#| label: 01-01
#| eval: false

install.packages("tidyverse")

```

We start by installing the "tidyverse" package, which includes a suite of data manipulation and visualization packages.

```{r}
#| label: 01-02
#| message: false

library(tidyverse)
library(readxl)

```

Next, we load the necessary libraries, including "tidyverse" for data manipulation and "readxl" for reading Excel files.

## Fetch and load data

This step involves fetching and loading a dataset named "TCGA-CDR" from an online source.

We define the URL of the dataset and the local path where it will be saved. The **`download.file`** function is used to download the dataset and store it in the specified location.

After downloading, we read the dataset into R using the "readxl" package's **`read_excel`** function.

```{r}
#| label: 01-03
#| warning: false

url <- "https://drive.google.com/uc?export=download&id=1pqlKgEo8foNOx5L_ULJfundKBX4-NNti"
path <- "../data/_raw/full_dataset.xlsx"

download.file(url = url, destfile = path, mode = "wb")

TCGA <- read_excel(path)
```

## Save in "data"

Finally, we save the loaded dataset in a tab-separated values (TSV) format for further analysis. The dataset is saved in the "data" directory as "01_dat_load.tsv" for easy access and future use in our analysis.

```{r}
#| label: 01-04

write_tsv(TCGA, "./../data/01_dat_load.tsv")
```
